gmail-mcp-agent-sdk/requirements.txt
# Official Model Context Protocol Python SDK
mcp>=1.2.0

# Gmail API
google-api-python-client==2.92.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0

# Utilities
python-dotenv==1.0.0
